Silence compiler warnings.
* common/openpgp-oid.c (map_openpgp_pk_to_gcry): Use cast for enum
conversion.
* dirmngr/dns-stuff.c (get_dns_srv): Use explicit conversion from
int to float.
* sm/gpgsm.c (parse_keyserver_line): Initialize ERR.

+++ dirmngr/dns-stuff.c
@@ -2196,7 +2196,7 @@ get_dns_srv (ctrl_t ctrl,
           (*list)[j].run_count=prio_count;
         }
 
-      chose=prio_count*rand()/RAND_MAX;
+      chose=prio_count*rand()/(float)RAND_MAX;
 
       for (j=i;j<srvcount && (*list)[i].priority==(*list)[j].priority;j++)
         {
